src/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.8
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-filter==22.1
django-phonenumber-field==5.0.0
django-split-settings==1.2.0
django-stubs-ext==0.5.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-builtins==1.5.3
flake8-comprehensions==3.8.0
flake8-eradicate==1.2.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.0.1
pytest-asyncio==0.18.3
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-flake8==1.1.1
pytest-forked==1.4.0
pytest-lazy-fixture
pytest-mock==3.7.0
pytest-xdist==2.5.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1